Extract CupertinoSwitchListTile and disable inkwell
